PDFBOX-2959: remove unneeded unboxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1701502 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/font/PDType3Font.java
@@ -131,7 +131,7 @@ public class PDType3Font extends PDSimpleFont
         int lastChar = dict.getInt(COSName.LAST_CHAR, -1);
         if (getWidths().size() > 0 && code >= firstChar && code <= lastChar)
         {
-            return getWidths().get(code - firstChar).floatValue();
+            return getWidths().get(code - firstChar);
         }
         else
         {
